changing the urls for the favorites to the users app
